Fix: invalid TinyDB query generation
This commit fixes one bug related to generating query in TinyDB format that is used to retrieve default hyperparameters for models and tasks. Search query (that is supposed to introduce constraints on model names and task types) was constructed using the wrong python API what was resulting in model type not being taken into account. This does not impact results published in the paper (since this implementation was added later, and back then default hyperparameters (which were exactly the same) were being retrieved using a different mechanism).