Adding dataset pre-requisites.
This commit adds a method in DatasetBuilder class to check for dataset pre-requisites before building any version of a dataset. Examples are - availability of optional dependencies, dataset files if they need to be downloaded manually, etc. All checks that existed in __init__ methods are moved to this new method in this commit. This also means that all dataset __init__ methods do not raise exceptions, and commands such as `dataset list` will work even when dataset pre-requisites are not met.