Making `openml` extra dependency. (#69)
XIME depends on `openml` library to download datasets hosted on OpenML site. This dependency is optional now, install it with `--extras openml`. The presence of this library is checked in `_check_pre_requisites` methods.
To implement this, a new class is added - DatasetPrerequisites - that will provide standard checks for dataset prerequisites.