Adding support for multiple parallel trials to use one GPU. (#83)
This commit changes the meaning of the `--gpu` command line argument parameter. In current implementation, it is a flag. When set, the Ray Tune scheduler assigns one trial to one GPU. This commit changes it to be an option with value being a GPU fraction - numerical value between 0 and 1 - that defines how much GPU a trial can potentially consume.

- When not provided, no GPUs will be used.
- When no value provided (`--gpu`), the value is set to 1 (this is backward compatible with current implementation).
- In all other cases, Ray Tune will use this value for scheduling multiple trials in parallel on a single GPU. For instance, with `--gpu 0.1` ten parallel trials can be scheduled on each GPU.