Adding new analysis function. (#84)
This commit adds new analysis function that computes some statistics of tree-based models (`XGBoost` and `RandomForest` from the `scikit-learn`). In addition, it adds information about model, input parameters and performance metric and represents it as a pandas data frame. This works for MLflow runs that are associated with the Ray Tune hyperparameter search experiments.