Setuptools version constraints for python < 3.11. (#102)
Older version of Ray that this project uses (`2.3.1` for python < `3.11`) are not compatible with the latest versions of the `setuptools` package (`ImportError: cannot import name 'packaging' from 'pkg_resources'`). The solution described [here](https://stackoverflow.com/questions/78604018/importerror-cannot-import-name-packaging-from-pkg-resources-when-trying-to)  is to use earlier version (s). This commit introduces a constraint for the `setuptools` package for python versions < `3.11` where there's a conflict with Ray.