Precision recall curve metric for binary classifiers. (#120)
This commit adds a new performance metric - precision recall (PR) curve. It works for binary classification models. It is serialized as two files:
 - Yaml file containing precision, recall and threshold arrays that were used to build the PR curve.
 - Image (png) file that is graphical representation of the curve with threshold-recall curve on the same chart.
The function that computes PR curve is decorated with a function decorator that ignores any exceptions raised during the process of computing the curve. This decorator (ignore_exceptions) is also part of this commit and can be used for other functions too.