PR curve looks differen on different systems. (#124)
This PR fixes a "bug" that was resulting in PR curve looking differently on different machines. This was due to using scikit-learn's `PrecisionRecallDisplay` class that internally (starting some version) updates passed matplotlib's axes. So different versions of scikit-learn resulted in different charts. This commit removes dependency on that class and plots PR by itself.