Correct dataset serialization.
Dataset.save method not checks for certain files that already present and that can impact subsequent data loading calls. If such files found, and overwrite is set to False, and exception is raised.
This commit also fixes the bug when a metadata file is always overwritten, and split files only overwritten when they do not present.